+++ src/main/java/qa/qcri/iyas/feature/similarity/CosineBowSimilarity.java
@@ -25,7 +25,7 @@ import com.google.common.base.Joiner;
 import cc.mallet.types.FeatureVector;
 import cc.mallet.types.NormalizedDotProductMetric;
 import qa.qcri.iyas.data.preprocessing.BowProvider;
-import qa.qcri.iyas.data.tree.nodes.RichNode;
+import qa.qcri.iyas.feature.tree.node.RichNode;
 
 
 /**
