+++ src/main/java/qa/qcri/iyas/feature/similarity/LongestCommonSubstringComparatorSimilarity.java
@@ -23,7 +23,8 @@ import com.google.common.base.Joiner;
 
 import de.tudarmstadt.ukp.similarity.algorithms.api.SimilarityException;
 import de.tudarmstadt.ukp.similarity.algorithms.lexical.string.LongestCommonSubstringComparator;
-import qa.qcri.iyas.data.tree.nodes.RichNode;
+import qa.qcri.iyas.feature.tree.node.RichNode;
+
 
 
 /**
