+++ src/main/java/qa/qcri/iyas/feature/similarity/WordNGramContainmentMeasureSimilarity.java
@@ -27,7 +27,7 @@ import com.google.common.base.Joiner;
 import de.tudarmstadt.ukp.similarity.algorithms.api.SimilarityException;
 import de.tudarmstadt.ukp.similarity.algorithms.lexical.ngrams.WordNGramContainmentMeasure;
 import de.tudarmstadt.ukp.similarity.algorithms.lexical.ngrams.WordNGramJaccardMeasure;
-import qa.qcri.iyas.data.tree.nodes.RichNode;
+import qa.qcri.iyas.feature.tree.node.RichNode;
 
 
 /**
