+++ src/main/java/qa/qcri/iyas/feature/similarity/WordNGramJaccardMeasureSimilarity.java
@@ -26,7 +26,8 @@ import com.google.common.base.Joiner;
 
 import de.tudarmstadt.ukp.similarity.algorithms.api.SimilarityException;
 import de.tudarmstadt.ukp.similarity.algorithms.lexical.ngrams.WordNGramJaccardMeasure;
-import qa.qcri.iyas.data.tree.nodes.RichNode;
+import qa.qcri.iyas.feature.tree.node.RichNode;
+
 
 
 /**
